libutt::api curve points can now be de/serialized as byte vectors to be consistent with other data. In this case we need to pass S2 from the server to the client which is a curve point

+++ utt/utt-common-api/include/utt-common-api/CommonApi.hpp
@@ -34,7 +34,7 @@ using UserRegistrationInput = std::vector<uint8_t>;
 using RegistrationSig = std::vector<uint8_t>;
 
 /// @brief A system generated part of the user's PRF key
-using S2 = std::vector<uint64_t>;
+using S2 = std::vector<uint8_t>;
 
 /// @brief The privacy budget determines how much in value can a user transfer anonymously to other users.
 using PrivacyBudget = std::vector<uint8_t>;
